gcp: add config option to impersonate service account
Impersonation allows to delegate to another service account, for
instance, when service account used in a build environment doesn't have
access to write to GCS.

Follows `--impersonate-service-account` as in gcloud nested under `gcp`
section in config. The only way to access config now is to set
`GCP_IMPERSONATE_SERVICE_ACCOUNT` environment variable.

This option is experimental and can change in the future.